+++ doc/source/changes.rst
@@ -14,6 +14,7 @@ New in 0.24
   :module:`mne` and are not cached any more. This can lead to small changes
   in results due to improved numerical precision. Delete old files to free up
   space with ``mne_experiment.rm('bem-sol-file', subject='*')``.
+* :meth:`MneExperiment.make_report_coreg` method.
 
 
 New in 0.23